backend/requirements.txt
requests==2.26.0
Django==3.2
djangorestframework==3.12.4
djoser==2.1.0
pytest-pythonpath==0.7.3
Pillow==11.0.0
django-filter==23.1
python-dotenv==1.0.1
psycopg2-binary==2.9.3
psycopg2==2.9.10